working on Shoping Cart

+++ Proj.Models/ViewModel/ShoppingCartVM.cs
@@ -9,7 +9,7 @@ namespace Proj.Models.ViewModel
     public class ShoppingCartVM
     {
         public IEnumerable<ShoppingCart> ShoppingCartList { get; set; }
-        public float OrderTotal { get; set; }
+        public double OrderTotal { get; set; }
 
     }
 }
